requirements.txt
flask
python-digitalocean
requests